# openapi.json
{"openapi":"3.1.0","info":{"title":"ClearanceHub API","version":"0.1.0"},"paths":{"/auth/request-otp":{"post":{"tags":["auth"],"summary":"Request Otp Endpoint","operationId":"request_otp_endpoint_auth_request_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOTP"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/verify-otp":{"post":{"tags":["auth"],"summary":"Verify Otp Endpoint","operationId":"verify_otp_endpoint_auth_verify_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOTP"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/card":{"post":{"tags":["requests"],"summary":"Create Card Request Endpoint","operationId":"create_card_request_endpoint_requests_card_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardRequestCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/access":{"post":{"tags":["requests"],"summary":"Create Access Request Endpoint","operationId":"create_access_request_endpoint_requests_access_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermitRequestCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{request_id}":{"get":{"tags":["requests"],"summary":"Get Request Detail Endpoint","operationId":"get_request_detail_endpoint_requests__request_id__get","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests":{"get":{"tags":["requests"],"summary":"List Requests Endpoint","operationId":"list_requests_endpoint_requests_get","parameters":[{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"From"}},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"To"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{request_id}/approve":{"post":{"tags":["approvals"],"summary":"Approve Request Endpoint","operationId":"approve_request_endpoint_requests__request_id__approve_post","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{request_id}/reject":{"post":{"tags":["approvals"],"summary":"Reject Request Endpoint","operationId":"reject_request_endpoint_requests__request_id__reject_post","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{request_id}/complete":{"post":{"tags":["approvals"],"summary":"Complete Request Endpoint","operationId":"complete_request_endpoint_requests__request_id__complete_post","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/requests/excel":{"get":{"tags":["reports"],"summary":"Export Requests Excel","operationId":"export_requests_excel_reports_requests_excel_get","parameters":[{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"From"}},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"To"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employees/{employee_id}":{"get":{"tags":["lookup"],"summary":"Get Employee","operationId":"get_employee_employees__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/areas":{"get":{"tags":["lookup"],"summary":"List Areas","operationId":"list_areas_areas_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AreaRead"},"title":"Response List Areas Areas Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ApprovalTimelineItem":{"properties":{"role":{"type":"string","title":"Role"},"employee_id":{"type":"string","title":"Employee Id"},"action":{"type":"string","title":"Action"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["role","employee_id","action","updated_at"],"title":"ApprovalTimelineItem"},"AreaRead":{"properties":{"area_id":{"type":"integer","title":"Area Id"},"area_name":{"type":"string","title":"Area Name"}},"type":"object","required":["area_id","area_name"],"title":"AreaRead"},"CardRequestCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"request_type":{"type":"string","enum":["NEW","RENEW","REPLACE_LOST"],"title":"Request Type"},"request_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Reason"},"photo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Url"}},"type":"object","required":["employee_id","request_type"],"title":"CardRequestCreate"},"CardRequestRead":{"properties":{"id":{"type":"integer","title":"Id"},"employee_id":{"type":"string","title":"Employee Id"},"submitted_by_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Submitted By Employee Id"},"request_date":{"type":"string","format":"date-time","title":"Request Date"},"request_type":{"type":"string","title":"Request Type"},"request_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Reason"},"photo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Url"},"manager_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Manager Employee Id"},"manager_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Manager Updated At"},"security_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Security Employee Id"},"security_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Security Updated At"},"printing_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Printing Employee Id"},"printing_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Printing Updated At"},"rejection_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rejection Reason"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","employee_id","request_date","request_type","status","created_at"],"title":"CardRequestRead"},"EmployeeRead":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"name_ar":{"type":"string","title":"Name Ar"},"name_en":{"type":"string","title":"Name En"},"job_title":{"type":"string","title":"Job Title"},"department_id":{"type":"integer","title":"Department Id"},"department_name":{"type":"string","title":"Department Name"},"account_status":{"type":"string","title":"Account Status"}},"type":"object","required":["employee_id","name_ar","name_en","job_title","department_id","department_name","account_status"],"title":"EmployeeRead"},"EmployeeSnapshot":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"name_ar":{"type":"string","title":"Name Ar"},"name_en":{"type":"string","title":"Name En"},"job_title":{"type":"string","title":"Job Title"},"department_id":{"type":"integer","title":"Department Id"},"department_name":{"type":"string","title":"Department Name"},"account_status":{"type":"string","title":"Account Status"},"nationality_ar":{"type":"string","title":"Nationality Ar"},"nationality_en":{"type":"string","title":"Nationality En"}},"type":"object","required":["employee_id","name_ar","name_en","job_title","department_id","department_name","account_status","nationality_ar","nationality_en"],"title":"EmployeeSnapshot"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PermitRequestCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"area_ids":{"items":{"type":"integer"},"type":"array","minItems":1,"title":"Area Ids"},"request_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Reason"}},"type":"object","required":["employee_id","area_ids"],"title":"PermitRequestCreate"},"PermitRequestRead":{"properties":{"id":{"type":"integer","title":"Id"},"employee_id":{"type":"string","title":"Employee Id"},"submitted_by_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Submitted By Employee Id"},"request_date":{"type":"string","format":"date-time","title":"Request Date"},"request_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Reason"},"manager_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Manager Employee Id"},"manager_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Manager Updated At"},"security_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Security Employee Id"},"security_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Security Updated At"},"printing_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Printing Employee Id"},"printing_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Printing Updated At"},"rejection_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rejection Reason"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"area_ids":{"items":{"type":"integer"},"type":"array","title":"Area Ids"},"areas":{"items":{"$ref":"#/components/schemas/AreaRead"},"type":"array","title":"Areas"}},"type":"object","required":["id","employee_id","request_date","status","created_at"],"title":"PermitRequestRead"},"RejectRequest":{"properties":{"reason":{"type":"string","title":"Reason"}},"type":"object","required":["reason"],"title":"RejectRequest"},"RequestDetailResponse":{"properties":{"request_type":{"type":"string","enum":["CARD","ACCESS"],"title":"Request Type"},"employee":{"$ref":"#/components/schemas/EmployeeSnapshot"},"card_request":{"anyOf":[{"$ref":"#/components/schemas/CardRequestRead"},{"type":"null"}]},"permit_request":{"anyOf":[{"$ref":"#/components/schemas/PermitRequestRead"},{"type":"null"}]},"approvals":{"items":{"$ref":"#/components/schemas/ApprovalTimelineItem"},"type":"array","title":"Approvals"}},"type":"object","required":["request_type","employee"],"title":"RequestDetailResponse"},"RequestListItem":{"properties":{"id":{"type":"integer","title":"Id"},"request_type":{"type":"string","enum":["CARD","ACCESS"],"title":"Request Type"},"employee_id":{"type":"string","title":"Employee Id"},"submitted_by_employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Submitted By Employee Id"},"status":{"type":"string","title":"Status"},"request_date":{"type":"string","format":"date-time","title":"Request Date"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"card_request_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Card Request Type"}},"type":"object","required":["id","request_type","employee_id","status","request_date","created_at"],"title":"RequestListItem"},"RequestListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/RequestListItem"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"RequestListResponse"},"RequestOTP":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"RequestOTP"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"TokenResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyOTP":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["email","otp"],"title":"VerifyOTP"}}}}